Move TxButton back to 123code (for the time being)
